refactor!: Removed some unused classes + Reorganized packages.
