+++ src/main/java/com/gersonfaneto/yams/controllers/CreatePaymentController.java
@@ -5,12 +5,12 @@ import java.io.IOException;
 import com.gersonfaneto.yams.App;
 import com.gersonfaneto.yams.dao.DAO;
 import com.gersonfaneto.yams.models.billing.invoice.Invoice;
-import com.gersonfaneto.yams.models.billing.payments.Payment;
-import com.gersonfaneto.yams.models.billing.payments.PaymentMethod;
+import com.gersonfaneto.yams.models.billing.payment.Payment;
+import com.gersonfaneto.yams.models.billing.payment.PaymentMethod;
 import com.gersonfaneto.yams.models.entities.client.Client;
 import com.gersonfaneto.yams.models.entities.technician.Technician;
-import com.gersonfaneto.yams.models.orders.work.WorkOrder;
-import com.gersonfaneto.yams.models.orders.work.WorkOrderState;
+import com.gersonfaneto.yams.models.services.order.WorkOrder;
+import com.gersonfaneto.yams.models.services.order.WorkOrderState;
 import com.gersonfaneto.yams.utils.TypeParser;
 import com.gersonfaneto.yams.views.components.PaymentsListComponent;
