+++ src/main/java/com/gersonfaneto/yams/controllers/InvoiceDetailsController.java
@@ -6,8 +6,8 @@ import java.util.List;
 import com.gersonfaneto.yams.App;
 import com.gersonfaneto.yams.dao.DAO;
 import com.gersonfaneto.yams.models.billing.invoice.Invoice;
-import com.gersonfaneto.yams.models.billing.payments.Payment;
-import com.gersonfaneto.yams.models.billing.payments.PaymentMethod;
+import com.gersonfaneto.yams.models.billing.payment.Payment;
+import com.gersonfaneto.yams.models.billing.payment.PaymentMethod;
 import com.gersonfaneto.yams.models.entities.client.Client;
 import com.gersonfaneto.yams.models.entities.user.UserType;
 import com.gersonfaneto.yams.utils.TypeParser;
