+++ src/main/java/com/gersonfaneto/yams/dao/billing/payment/PaymentMemoryDAO.java
@@ -1,7 +1,7 @@
-package com.gersonfaneto.yams.dao.billing.paymet;
+package com.gersonfaneto.yams.dao.billing.payment;
 
 import com.gersonfaneto.yams.dao.CRUD;
-import com.gersonfaneto.yams.models.billing.payments.Payment;
+import com.gersonfaneto.yams.models.billing.payment.Payment;
 import com.gersonfaneto.yams.utils.Generators;
 import java.util.HashMap;
 import java.util.List;
