+++ src/main/java/com/gersonfaneto/yams/dao/services/order/WorkOrderCRUD.java
@@ -1,7 +1,8 @@
-package com.gersonfaneto.yams.dao.orders.work;
+package com.gersonfaneto.yams.dao.services.order;
 
 import com.gersonfaneto.yams.dao.CRUD;
-import com.gersonfaneto.yams.models.orders.work.WorkOrder;
+import com.gersonfaneto.yams.models.services.order.WorkOrder;
+
 import java.util.List;
 
 /**
