+++ src/main/java/com/gersonfaneto/yams/dao/services/order/WorkOrderMemoryDAO.java
@@ -1,7 +1,7 @@
-package com.gersonfaneto.yams.dao.orders.work;
+package com.gersonfaneto.yams.dao.services.order;
 
 import com.gersonfaneto.yams.dao.CRUD;
-import com.gersonfaneto.yams.models.orders.work.WorkOrder;
+import com.gersonfaneto.yams.models.services.order.WorkOrder;
 import com.gersonfaneto.yams.utils.Generators;
 import java.util.HashMap;
 import java.util.List;
