+++ src/main/java/com/gersonfaneto/yams/utils/TypeParser.java
@@ -1,9 +1,9 @@
 package com.gersonfaneto.yams.utils;
 
-import com.gersonfaneto.yams.models.billing.payments.PaymentMethod;
+import com.gersonfaneto.yams.models.billing.payment.PaymentMethod;
 import com.gersonfaneto.yams.models.entities.user.UserType;
-import com.gersonfaneto.yams.models.orders.work.WorkOrderState;
-import com.gersonfaneto.yams.models.services.ServiceType;
+import com.gersonfaneto.yams.models.services.order.WorkOrderState;
+import com.gersonfaneto.yams.models.services.service.ServiceType;
 import com.gersonfaneto.yams.models.stock.ComponentType;
 
 public abstract class TypeParser {
