+++ src/test/java/com/gersonfaneto/yams/dao/services/service/ServiceDAOTest.java
@@ -1,9 +1,9 @@
-package com.gersonfaneto.yams.dao.services;
+package com.gersonfaneto.yams.dao.services.service;
 
 import com.gersonfaneto.yams.dao.DAO;
 import com.gersonfaneto.yams.dao.Persist;
-import com.gersonfaneto.yams.models.services.Service;
-import com.gersonfaneto.yams.models.services.ServiceType;
+import com.gersonfaneto.yams.models.services.service.Service;
+import com.gersonfaneto.yams.models.services.service.ServiceType;
 import com.gersonfaneto.yams.utils.Generators;
 import java.io.File;
 import java.util.List;
